test(core): split option combination tests
