+++ packages/core/test/core/option-combinations/core-options/splitting.ts
@@ -1,0 +1,36 @@
+import {
+  afterAll,
+  beforeAll,
+  build,
+  createOptionCombinationFixture,
+  describe,
+  expect,
+  join,
+  removeOptionCombinationFixture,
+  test,
+  writeFileSync,
+} from '../helpers';
+
+describe('옵션 조합 통합 테스트 - core options - splitting', () => {
+  let dir: string;
+
+  beforeAll(() => {
+    dir = createOptionCombinationFixture();
+  });
+
+  afterAll(() => {
+    removeOptionCombinationFixture(dir);
+  });
+
+  test('splitting + entryNames + chunkNames 조합', async () => {
+    writeFileSync(join(dir, 'dyn-entry.ts'), 'export const lazy = () => import("./lib");');
+    const result = await build({
+      entryPoints: [join(dir, 'dyn-entry.ts')],
+      splitting: true,
+      entryNames: '[name]',
+      chunkNames: 'chunks/[name]-[hash]',
+    });
+    expect(result.errors.length).toBe(0);
+    expect(result.outputFiles.length).toBeGreaterThanOrEqual(2);
+  });
+});
